requirements.txt
requests
shapely==1.6.4
polyline==1.3.2
django==2.0.1
pillow==5.0.0
regex==2018.01.10
